[FLINK-4282]Add Offset Parameter to WindowAssigners
Reformat code, move tests to suitable place and revert unrelated code change
