+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/windowing/assigners/SlidingProcessingTimeWindows.java
@@ -51,7 +51,7 @@ public class SlidingProcessingTimeWindows extends WindowAssigner<Object, TimeWin
 
 	private final long slide;
 
-	private SlidingProcessingTimeWindows(long size,long slide,long offset){
+	private SlidingProcessingTimeWindows(long size, long slide, long offset){
 		this.size = size;
 		this.slide = slide;
 		this.offset = offset;
@@ -61,7 +61,7 @@ public class SlidingProcessingTimeWindows extends WindowAssigner<Object, TimeWin
 	public Collection<TimeWindow> assignWindows(Object element, long timestamp, WindowAssignerContext context) {
 		timestamp = context.getCurrentProcessingTime();
 		List<TimeWindow> windows = new ArrayList<>((int) (size / slide));
-		long lastStart = TimeWindow.getWindowStartWithOffset(timestamp,offset,slide);
+		long lastStart = TimeWindow.getWindowStartWithOffset(timestamp, offset, slide);
 		for (long start = lastStart;
 			start > timestamp - size;
 			start -= slide) {
@@ -97,7 +97,7 @@ public class SlidingProcessingTimeWindows extends WindowAssigner<Object, TimeWin
 	 * @return The time policy.
 	 */
 	public static SlidingProcessingTimeWindows of(Time size, Time slide) {
-		return new SlidingProcessingTimeWindows(size.toMilliseconds(), slide.toMilliseconds(),0);
+		return new SlidingProcessingTimeWindows(size.toMilliseconds(), slide.toMilliseconds(), 0);
 	}
 
 	/**
@@ -120,7 +120,7 @@ public class SlidingProcessingTimeWindows extends WindowAssigner<Object, TimeWin
 	 * @param offset The offset which window start would be shifted by.
 	 * @return The time policy.
 	 */
-	public static SlidingProcessingTimeWindows of(Time size, Time slide,Time offset) {
+	public static SlidingProcessingTimeWindows of(Time size, Time slide, Time offset) {
 		return new SlidingProcessingTimeWindows(size.toMilliseconds(), slide.toMilliseconds(),
 			offset.toMilliseconds() % slide.toMilliseconds());
 	}
