+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/windowing/assigners/SlidingTimeWindows.java
@@ -33,7 +33,7 @@ public class SlidingTimeWindows extends SlidingEventTimeWindows {
 	private static final long serialVersionUID = 1L;
 
 	private SlidingTimeWindows(long size, long slide) {
-		super(size, slide,0);
+		super(size, slide, 0);
 	}
 
 	/**
